+++ commercetools/commercetools-graphql-api/src/integrationTest/java/com/commercetools/graphql/CommercetoolsTestUtils.java
