+++ commercetools/commercetools-graphql-api/src/main/resources/graphql/schema.graphqls
