+++ settings.gradle
@@ -5,7 +5,7 @@ include 'commercetools:commercetools-okhttp-client4'
 include 'rmf:okhttp-client'
 include 'commercetools:commercetools-http-client'
 include 'commercetools:commercetools-monitoring-newrelic'
-include 'commercetools:commercetools-graphql'
+include 'commercetools:commercetools-graphql-api'
 include 'commercetools:commercetools-sdk-java-api'
 include 'commercetools:commercetools-sdk-java-importapi'
 include 'commercetools:commercetools-sdk-java-ml'
